Add files via upload
This is an install script only for SkyTrak's app. This will allow you to have a separate install wherever you choose. 

This asks where you want to install the app, then downloads it for you and installs it in that location. Will not impede an existing install, unless you select the same location your SkyTrak app is installed.